requirements.txt
mysql-connector
pn532pi
rpi-lgpio
pyserial
